# packages/db-api/openapi.yaml
components:
  securitySchemes:
    cookieAuth:
      in: cookie
      name: session
      type: apiKey
info:
  description:
    'The HR DB API is a RESTful API that provides access to the HR database.


    API possible responses: https://developer.mozilla.org/en-US/docs/Web/HTTP/Status#client_error_responses


    Informational responses (100 - 199)


    Successful responses (200 - 299)


    Redirection messages (300 - 399)


    Client error responses (400 - 499)


    Server error responses (500 - 599)

    '
  title: HR DB API,
  version: apiVersion,
openapi: 3.1.0
paths:
...
  /groups/{id}:
    delete:
      parameters:
        - description: The ID of the group to delete
          in: path
          name: id
          required: true
          schema:
            type: integer
      responses:
        200:
          content:
            application/json:
              schema:
                properties:
                  message:
                    example: Group 1 deleted successfully
                    type: string
                type: object
          description: Group deleted successfully
      summary: Delete a group by ID
      tags:
        - Groups
    get:
      parameters:
        - description: The ID of the group
          in: path
          name: id
          required: true
          schema:
            type: integer
      responses:
        200:
          content:
            application/json:
              schema:
                properties:
                  description:
                    example: Group focused on development tasks.
                    type: string
                  groupUsers:
                    items:
                      properties:
                        user:
                          properties:
                            email:
                              example: john.doe@example.com
                              type: string
                            firstName:
                              example: John
                              type: string
                            lastName:
                              example: Doe
                              type: string
                          type: object
                        userId:
                          example: 101
                          type: integer
                      type: object
                    type: array
                  id:
                    example: 1
                    type: integer
                  name:
                    example: Development Group
                    type: string
                type: object
          description: Group retrieved successfully
        404:
          content:
            application/json:
              schema:
                properties:
                  message:
                    example: Group 1 not found
                    type: string
                type: object
          description: Group not found
      summary: Retrieve a specific group by ID
      tags:
        - Groups
    put:
      parameters:
        - description: The ID of the group to update
          in: path
          name: id
          required: true
          schema:
            type: integer
      requestBody:
        content:
          application/json:
            schema:
              properties:
                description:
                  example: Group focused on development tasks.
                  type: string
                name:
                  example: Development Group
                  type: string
                location:
                  properties:
                    id:
                      example: 1
                      type: integer
                    name:
                      example: New York Office
                      type: string
                  type: object
              type: object
        required: true
      responses:
        200:
          content:
            application/json:
              schema:
                properties:
                  message:
                    example: Group 1 updated successfully
                    type: string
                type: object
          description: Group updated successfully
      summary: Update a group by ID
      tags:
        - Groups
...
security:
  - cookieAuth: []
servers:
  - description: Development server
    url: http://localhost:5000/api/v1